# Misc perf improvements