--- a/src/plot_api/plot_api.js
+++ b/src/plot_api/plot_api.js
@@ -189,9 +189,7 @@ Plotly.plot = function(gd, data, layout, config) {
         }
 
         return Lib.syncOrAsync([
-            subroutines.layoutStyles,
-            drawAxes,
-            initInteractions
+            subroutines.layoutStyles
         ], gd);
     }