--- a/src/plot_api/plot_api.js
+++ b/src/plot_api/plot_api.js
@@ -368,7 +368,6 @@ Plotly.plot = function(gd, data, layout, config) {
         drawFramework,
         marginPushers,
         marginPushersAgain,
-        initInteractions,
         positionAndAutorange,
         subroutines.layoutStyles,
         drawAxes,