… src/components/legend/style.js
… src/plots/gl2d/scene2d.js
… src/plots/gl3d/scene.js
… src/traces/bar/plot.js
… src/traces/bar/style.js
… src/traces/box/style.js
… src/traces/pie/plot.js
… src/traces/sankey/plot.js
… src/traces/scatter/plot.js
… src/traces/scatter/style.js
… src/traces/scattergeo/plot.js
… test/jasmine/tests/bar_test.js
… test/jasmine/tests/cartesian_interact_test.js
… test/jasmine/tests/dragelement_test.js