A bunch of perf improvements. The biggest one is changing how we cache `Drawing.bBox` results https://github.com/plotly/plotly.js/commit/ea883e6912c216bc8243d0f18b67b098233df0e7 - instead of having an element keep track of its own caching, I changed it to just using the contents of the node (its `innerHTML` plus its `style` and `text-anchor` - I'm 90% sure this is sufficient to determine the bbox but needs a bit more testing). This means we're not always invalidating the cache, whether accidentally or out of an abundance of caution, and in fact even unrelated elements that have the same contents and styling can use the same bbox (such as matching labels on different axes).

Another fairly significant one is removing an unnecessary double-draw of ticks and grids during `Plotly.plot` https://github.com/plotly/plotly.js/commit/b4a2f4d581c14c92f1b9733407b1f3a9e0b27aa8 - no idea where that came from but it doesn't seem to cause any problems to pare that down.

Two smaller ones:
- don't draw placeholder titles that we're just going to delete https://github.com/plotly/plotly.js/commit/c02204f13f6508b26039ea14b54944d4352ff7c8
- remove `Lib.getPlotDiv` https://github.com/plotly/plotly.js/commit/032cf7c191cd10000ccc7a77ca3d0cb03edbbeef - this was by far the most intrusive change (I had to pass `gd` around to a bunch of places that didn't have it before), for the smallest benefit, but it turns out `instanceof HTMLElement` (which is actually in `Lib.isPlotDiv` which I *didn't* get rid of... but that one calls it once and `getPlotDiv` called it in a loop) could be kind of slow.

Also I added a perf tester `timeit` to the global scope in the test dashboard, just for convenience.

I'd like to get #1767 merged before finishing this so I can bring the test improvements in here.