@monfera 2017-06-09
Looks great! 

- I've often seen double or worse rendering in imperative code, hard to automatically catch, I guess this double render was not intended
- bounding box calc is in good part expensive because the layoutung must ve synchronously awaited, so I'd guess avoiding bbox calc on identical tick text is even bigger than avoiding repeat calls
- caching is best if global, but risk is mem leak