@alexcjohnson 2017-06-11
Some profiling on small charts: All results from Chrome on my mac with no profiles running. Running profiles adds ~5% to these times with no screenshots, ~10% with screenshots (interestingly the profiling penalty seems bigger even *fractionally* before the change). All results are averaging ~10 runs of `Plotly.newPlot` on the same data repeatedly, throwing away the first one. This gives extra weight to the caching improvements, but as my interest is mainly in interactivity performance that's what I want.

using mock `category_dtick_3` (small line chart with a linear and a category axis):
- before this branch: 144ms
- after this branch: 63ms

using a chart with ~30 variable-size bubbles, gradient fill, long category names:
- before this branch: 444ms
- after this branch: 270ms